pyproject.toml
[project]
name = "blog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-fasthtml>=0.12.1",
    "python-dotenv>=1.0.1",
    "lucide-fasthtml>=0.0.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
